cli: Move CliOutput to picard/cli/output.py as shared base
The colorized terminal output class (print, success, error, warning,
color helpers, semantic display methods, yesno prompts) is generic
infrastructure useful for any CLI subcommand, not just plugins.

- picard/cli/output.py: CliOutput class (the canonical location)
- picard/plugin3/output.py: Re-exports CliOutput as PluginOutput
  for full backward compatibility (all existing imports unchanged)

Future subcommands (profiles, config, etc.) can import CliOutput
directly from picard.cli.output.
